WIP
Start migrating and refactoring actual count
